ultimos cambios -algo anda-

+++ tp3/entregable/tp3-bundle.v1/src/game.h
@@ -15,7 +15,7 @@ typedef enum direccion_e { IZQ = 0xAAA, DER = 0x441, ADE = 0x83D, ATR = 0x732 } 
 
 void game_jugador_mover(unsigned int jugador, unsigned int value);
 
-void game_lanzar_zombi(unsigned int jugador);
+void game_lanzar_zombi(u8 jugador);
 
 void game_move_current_zombi(direccion dir);
